requirements.txt
aiohttp
asyncio
discord.py
sqlite3